Move filter to a subcommand
